requirements.txt
absl-py==2.0.0
astunparse==1.6.3
biopython==1.81
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distinctipy==1.2.3
docopt==0.6.2
flatbuffers==23.5.26
fonttools==4.44.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
joblib==1.2.0
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
ml-dtypes==0.2.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipreqs==0.4.13
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.66.1
trimesh==4.0.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wget==3.2
wrapt==1.14.1
yarg==0.1.9
